cloud-run/api/requirements.txt
fastapi
uvicorn
requests
google-cloud-firestore
python-dotenv
aiofiles
